% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{api_key}
\alias{api_key}
\title{Return the user ID from the PTV_API_KEY environment variable}
\usage{
api_key()
}
\value{
Character. The value of the PTV_API_KEY environment variable.
}
\description{
Currently this package only supports obtaining the API key either directly
as an argument to functions, or from the PTV_API_KEY environment variable.
This function will error if the PTV_API_KEY environment variable is not set.
A message will also be displayed to the user, explaining how to obtain
authentication details from PTV. This message will be displayed only once
per session.
}
\keyword{internal}
