% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{assert_correct_attributes}
\alias{assert_correct_attributes}
\title{Assert that the API has returned the expected attributes}
\usage{
assert_correct_attributes(received_attributes, expected_attributes)
}
\arguments{
\item{received_attributes}{A character vector of attributes, in order.}

\item{expected_attributes}{A character vector of expected attributes, in
order.}
}
\value{
An error if the column names are not as expected.
}
\description{
The attributes returned by the API calls should be follow the API schema.
This function compares received attributes against a vector of expected
attributes, and returns an error if the two do not match. Unfortunately,
there is no easy fix for this error: the package developer(s) must be
notified, so that they can align the functions against the API schema.
}
\keyword{internal}
