% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert-correct-colnames.R
\name{assert_correct_colnames}
\alias{assert_correct_colnames}
\title{Assert that the API has returned the expected attributes}
\usage{
assert_correct_colnames(data, expected_colnames)
}
\arguments{
\item{data}{A tibble parsed from the content returned by a PTV API call.}

\item{expected_colnames}{A character vector of expected column names, in
order.}
}
\value{
An error if the column names are not as expected.
}
\description{
The attributes returned by the API calls are usually converted to column
names in a tibble. This function compares that tibble against a vector of
expected column names, and returns an error if the two do not match.
Unfortunately, there is no easy fix for this error: the package developer(s)
must be notified, so that they can align the functions against the API
schema.
}
\keyword{internal}
