% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{convert_to_melbourne_time}
\alias{convert_to_melbourne_time}
\title{Convert a datetime returned by the PTV API into Melbourne time}
\usage{
convert_to_melbourne_time(datetime)
}
\arguments{
\item{datetime}{A datetime returned by the PTV API}
}
\value{
A datetime in the Melbourne timezone.
}
\description{
Convert a datetime returned by the PTV API into Melbourne time
}
\examples{
convert_to_melbourne_time("2020-05-09T06:38:47.3194196+00:00")
}
\keyword{internal}
