% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departures.R
\name{departure_to_tibble}
\alias{departure_to_tibble}
\title{Convert a single departure to a tibble}
\usage{
departure_to_tibble(departure)
}
\arguments{
\item{departure}{A departure, as a list, returned by the \code{departures} API
call.}
}
\value{
A tibble consisting of the following columns: \itemize{
\item \code{stop_id}
\item \code{route_id}
\item \code{run_id} (deprecated, use \code{run_ref} instead)
\item \code{run_ref}
\item \code{direction_id}
\item \code{disruption_ids}
\item \code{scheduled_departure}
\item \code{estimated_departure}
\item \code{at_platform}
\item \code{platform_number}
\item \code{flags}
\item \code{departure_sequence}
}
}
\description{
This function is designed to parse the content returned by the interior
steps of the \code{departures} functions.
}
\keyword{internal}
