% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disruptions.R
\name{disruption_information}
\alias{disruption_information}
\title{Stop information (metropolitan and V/Line stations only).}
\usage{
disruption_information(
  disruption_id,
  user_id = determine_user_id(),
  api_key = determine_api_key()
)
}
\arguments{
\item{user_id}{Integer or character. A user ID or devid provided by Public
Transport Victoria. See section "Obtaining API authentication details".}

\item{api_key}{Character. An API key, with dashes, provided by Public
Transport Victoria. See section "Obtaining API authentication details".}

\item{stop_id}{Integer disruption ID.}
}
\description{
This function can be used when integer disruption ID is already known. This
can be searched for with either the \code{disruptions_on_route} or
\code{disruptions_at_stop} functions.
}
\section{Swagger documentation}{

\url{http://timetableapi.ptv.vic.gov.au/swagger/ui/index#/Stops}
}

