% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disruptions.R
\name{disruption_to_tibble}
\alias{disruption_to_tibble}
\title{Convert a single disruptions to a tibble}
\usage{
disruption_to_tibble(disruption)
}
\arguments{
\item{disruption}{A disruption, as a list, returned by the \code{disruptions} API
call.}
}
\value{
A tibble with the following columns: \itemize{
\item \code{disruption_id}
\item \code{title}
\item \code{url}
\item \code{description}
\item \code{disruption_status}
\item \code{disruption_type}
\item \code{published_on}
\item \code{last_updated}
\item \code{from_date}
\item \code{to_date}
\item \code{routes}
\item \code{stops}
\item \code{colour}
\item \code{display_on_board}
\item \code{display_status}
}
}
\description{
This function is designed to parse the content returned by the interior
steps of the \code{disruptions_on_route} and \code{disruptions_at_stop} functions.
}
