% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disruptions.R
\name{disruptions}
\alias{disruptions}
\title{Stop information (metropolitan and V/Line stations only).}
\usage{
disruptions(
  disruption_id = NULL,
  route_types = NULL,
  disruption_modes = NULL,
  disruption_status = NULL,
  user_id = determine_user_id(),
  api_key = determine_api_key()
)
}
\arguments{
\item{disruption_id}{Integer. Optionally filter by a disruption ID.}

\item{route_types}{Optionally filter by a vector of route types. A route type
can be provided either as a non-negative integer code, or as a character:
"Tram", "Train", "Bus", "Vline" or "Night Bus". Character inputs are not
case-sensitive. Use the \code{route_types} function to extract a vector of all
route types.}

\item{disruption_modes}{Integer vector. Optionally filter by disruption
modes. For a full list of modes and their corresponding descriptions, use
the \code{description_modes()} function.}

\item{disruption_status}{Character. Can be used to filter to either "current"
or "planned" disruptions. Defaults to NULL, in which case no filter is
applied.}

\item{user_id}{Integer or character. A user ID or devid provided by Public
Transport Victoria. See section "Obtaining API authentication details".}

\item{api_key}{Character. An API key, with dashes, provided by Public
Transport Victoria. See section "Obtaining API authentication details".}
}
\description{
This function can be used when integer disruption ID is already known. This
can be searched for with either the \code{disruptions_on_route} or
\code{disruptions_at_stop} functions.
}
\section{Swagger documentation}{

\url{http://timetableapi.ptv.vic.gov.au/swagger/ui/index#/Stops}
}

