% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fare-estimate.R
\name{fare_estimate}
\alias{fare_estimate}
\title{Calculate a fare estimate between zones.}
\usage{
fare_estimate(
  min_zone,
  max_zone,
  journey_touch_on = NULL,
  journey_touch_off = NULL,
  journey_in_free_tram_zone = FALSE,
  travelled_route_types = NULL,
  user_id = determine_user_id(),
  api_key = determine_api_key()
)
}
\arguments{
\item{min_zone}{Integer. Minimum zone travelled through.}

\item{max_zone}{Integer. Maximum zone travelled through.}

\item{journey_touch_on}{POSIXct or Character. Optionally filter results to a
journey time. If providing a value to this variable, a value must also be
provided to \code{journey_touch_off}. Characters are automatically converted to
datetimes, and are assumed to be given as Melbourne time.}

\item{journey_touch_off}{POSIXct or Character. Optionally filter results to a
journey time. If providing a value to this variable, a value must also be
provided to \code{journey_touch_ofn}. Characters are automatically converted to
datetimes, and are assumed to be given as Melbourne time.}

\item{journey_in_free_tram_zone}{Boolean. Defaults to \code{FALSE}.}

\item{travelled_route_types}{Integer or character vector. Optionally filter
by a vector of route types. A route type can be provided either as a
non-negative integer code, or as a character: "Tram", "Train", "Bus",
"Vline" or "Night Bus". Character inputs are not case-sensitive. Use the
\code{route_types} function to extract a vector of all route types.}

\item{user_id}{Integer or character. A user ID or devid provided by Public
Transport Victoria. See section "Obtaining API authentication details".}

\item{api_key}{Character. An API key, with dashes, provided by Public
Transport Victoria. See section "Obtaining API authentication details".}
}
\value{
A tibble consisting of the following columns: \itemize{
\item \code{direction_id}
\item \code{direction_name},
\item \code{route_id}
\item \code{route_type}
\item \code{route_direction_description}
}
}
\description{
Calculate a fare estimate between zones.
}
