% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-request-url.R
\name{generate_request_url}
\alias{generate_request_url}
\title{Generate a URL with devid and signature}
\usage{
generate_request_url(
  request,
  user_id = determine_user_id(),
  api_key = determine_api_key()
)
}
\arguments{
\item{request}{A request or path for the API, eg. "routes".}

\item{user_id}{Integer or character. A user ID or devid provided by Public
Transport Victoria. See section "Obtaining API authentication details".}

\item{api_key}{Character. An API key, with dashes, provided by Public
Transport Victoria. See section "Obtaining API authentication details".}
}
\value{
A complete URL character that can be queried with httr
}
\description{
Generate a URL with devid and signature
}
\keyword{internal}
