% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-request-url.R
\name{generate_request_url_without_auth}
\alias{generate_request_url_without_auth}
\title{Generate a URL without devid and signature}
\usage{
generate_request_url_without_auth(request)
}
\arguments{
\item{request}{A request or path for the API, eg. "routes".}
}
\value{
A complete URL character that can be queried with httr
}
\description{
Generate a URL without devid and signature
}
\keyword{internal}
