% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopath.R
\name{geopath_to_tibble}
\alias{geopath_to_tibble}
\title{Convert a single geopath to a tibble}
\usage{
geopath_to_tibble(geopath)
}
\arguments{
\item{geopath}{A \code{geopaths} object, as a list}
}
\value{
A tibble of routes, with the following columns:
\itemize{
\item \code{direction_id}
\item \code{valid_from}
\item \code{valid_to}
\item \code{paths}
}
}
\description{
This function is designed to parse the \code{geopath} content returned by the
interior steps of some functions. If geopath data is requested, that content
will contain a list of \code{geopaths} for each route. This function is designed
to parse \emph{one} of those geopaths into a tibble.
}
\keyword{internal}
