% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url-manipulation.R
\name{make_url_friendly}
\alias{make_url_friendly}
\title{Convert an input to a form that can be used in a URL.}
\usage{
make_url_friendly(input)
}
\arguments{
\item{input}{Character or logical.}
}
\value{
Character.
}
\description{
Before a character can be used as part of a html, spaces must be converted to
"\%20". Logicals must be either "true" or "false", as characters.
}
\keyword{internal}
