% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{map_and_rbind}
\alias{map_and_rbind}
\title{Map and rbind a list of data frames}
\usage{
map_and_rbind(.x, .f, ...)
}
\arguments{
\item{.x}{A list of data frames or tibbles.}

\item{.f}{A function.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A data frame or tibble.
}
\description{
This function is a simple combination of \code{purrr::map()} and \code{purrr::reduce()}
using \code{rbind}. This differs from \code{purrr::map_dfr()}, which uses
\code{dplyr::map_dfr()} and therefore introduces \code{dplyr} as a dependency. If the
provided list is empty, then an empty tibble will be returned.
}
\keyword{internal}
