% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{need_api_authentication_details}
\alias{need_api_authentication_details}
\title{Message to display when API details are missing}
\usage{
need_api_authentication_details()
}
\value{
Character. A message explaining how to obtain and use authentication
details for the PTV API.
}
\description{
This function outputs a string that guides the user in obtaining a user ID
and API key so that they can access the PTV API. It sets an option so that
other functions can check if this message has already been displayed, to
avoid repetition.
}
\keyword{internal}
