% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlets.R
\name{outlet_to_tibble}
\alias{outlet_to_tibble}
\title{Convert a single outlet to a tibble}
\usage{
outlet_to_tibble(outlet)
}
\arguments{
\item{outlet}{An outlet, as a list, returned by the \code{outlets} API call.}
}
\value{
A tibble with the following columns: \itemize{
\item {\code{outlet_slid_spid}}
\item {\code{outlet_name}}
\item {\code{outlet_business}}
\item {\code{outlet_latitude}}
\item {\code{outlet_longitude}}
\item {\code{outlet_suburb}}
\item {\code{outlet_postcode}}
\item {\code{outlet_business_hour_mon}}
\item {\code{outlet_business_hour_tue}}
\item {\code{outlet_business_hour_wed}}
\item {\code{outlet_business_hour_thu}}
\item {\code{outlet_business_hour_fri}}
\item {\code{outlet_business_hour_sat}}
\item {\code{outlet_business_hour_sun}}
\item {\code{outlet_notes}}
}
}
\description{
This function is designed to parse the content returned by the interior
steps of the \code{outlets} and \code{outlets_nearby} functions.
}
\keyword{internal}
