% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlets.R
\name{outlets}
\alias{outlets}
\title{Information for a all outlets}
\usage{
outlets(user_id = determine_user_id(), api_key = determine_api_key())
}
\arguments{
\item{user_id}{Integer or character. A user ID or devid provided by Public
Transport Victoria. Refer to \code{?ptvapi} for more details.}

\item{api_key}{Character. An API key, with dashes, provided by Public
Transport Victoria. Refer to \code{?ptvapi} for more details.}
}
\value{
A tibble with the following columns: \itemize{
\item {\code{outlet_slid_spid}}
\item {\code{outlet_name}}
\item {\code{outlet_business}}
\item {\code{outlet_latitude}}
\item {\code{outlet_longitude}}
\item {\code{outlet_suburb}}
\item {\code{outlet_postcode}}
\item {\code{outlet_business_hour_mon}}
\item {\code{outlet_business_hour_tue}}
\item {\code{outlet_business_hour_wed}}
\item {\code{outlet_business_hour_thu}}
\item {\code{outlet_business_hour_fri}}
\item {\code{outlet_business_hour_sat}}
\item {\code{outlet_business_hour_sun}}
\item {\code{outlet_notes}}
}
}
\description{
Information for a all outlets
}
\details{
The business hours are reported as characters. Usually they take on
a format of "8.00AM - 10.00PM", but there variants such as "7.30AM - 11.00AM
and 1.30PM - 6.00PM". For days on which an outlet is closed, the opening
hours are usually reported as "CLOSED", but can also be an empty character.
Some opening hours are "24 Hours". These fields are also filled with missing
values and empty characters.
}
\examples{
\dontrun{
outlets()
}

}
