% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directions.R
\name{parse_directions_content}
\alias{parse_directions_content}
\title{Parse content of directions API call}
\usage{
parse_directions_content(directions_content)
}
\arguments{
\item{directions_content}{API response returned in the interior of a
directions function.}
}
\value{
A tibble consisting of the following columns: \itemize{
\item \code{direction_id}
\item \code{direction_name},
\item \code{route_id}
\item \code{route_type}
\item \code{route_direction_description}
}
}
\description{
This function is designed to parse the content
}
