% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fare-estimate.R
\name{parse_fare_estimate_content}
\alias{parse_fare_estimate_content}
\title{Parse content of fare estimates API call}
\usage{
parse_fare_estimate_content(fare_estimate_content)
}
\arguments{
\item{fare_estimate_content}{A direction, as a list, returned by the
\code{fare_estimate} API call.}
}
\value{
A data frame consisting of one row for each \code{passenger_type}, and the
following columns: \itemize{
\item \code{min_zone}
\item \code{max_zone}
\item \code{unique_zones}
\item \code{early_bird}
\item \code{free_tram_zone}
\item \code{weekend_journey}
\item \code{passenger_type}
\item \code{fare_2_hour_peak}
\item \code{fare_2_hour_off_peak}
\item \code{fare_daily_peak}
\item \code{fare_daily_off_peak}
\item \code{pass_7_days}
\item \code{pass_28_to_69_day_per_day}
\item \code{pass_70_plus_day_per_day}
\item \code{weekend_cap}
\item \code{holiday_cap}
}
}
\description{
This function is designed to parse the content returned by the interior
steps of the \code{fare_estimate} function.
}
\keyword{internal}
