% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url-manipulation.R
\name{prefix_base_url}
\alias{prefix_base_url}
\title{Prefix a string with the API base URL}
\usage{
prefix_base_url(string, and_version = FALSE)
}
\value{
character of form \code{{base_url}/{string}}
}
\description{
A trailing \code{/} will be added to the base URL if the input string does not
begin with a \code{/}.
}
\keyword{internal}
