% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url-manipulation.R
\name{prefix_base_url_and_version}
\alias{prefix_base_url_and_version}
\title{Prefix a string with the API base URL and version}
\usage{
prefix_base_url_and_version(string)
}
\value{
character of form \code{{base_url}/{version}/{string}}
}
\description{
A trailing \code{/} will be added to the base URL and version if the input string
does not begin with a \code{/}.
}
\keyword{internal}
