% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url-manipulation.R
\name{prefix_version}
\alias{prefix_version}
\title{Prefix a string with the API version}
\usage{
prefix_version(string)
}
\value{
character of form \verb{/\{version\}/\{string\}}
}
\description{
A trailing \code{/} will be added to the version if the input string does not
begin with a \code{/}.
}
\keyword{internal}
