% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptv-verbs.R
\name{process_response}
\alias{process_response}
\title{Process a raw httr response and return an object of class ptv_api}
\usage{
process_response(response, request_url_without_auth)
}
\arguments{
\item{response}{A raw response generated by the \code{httr} package}

\item{request_url_without_auth}{Character. The request \code{url}, without \code{devid}
and signature}
}
\value{
An S3 object of class ptv_api
}
\description{
This S3 object returned by this function contains three elements:
\itemize{
\item status_code, as an integer
\item request, which is the request URL without authentication details
\item content, the unparsed body of the response
}
}
\keyword{internal}
