% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{ptv_search}
\alias{ptv_search}
\title{Use a character term to search for routes, stops, and outlets.}
\usage{
ptv_search(
  search_term,
  user_id = determine_user_id(),
  api_key = determine_api_key()
)
}
\arguments{
\item{search_term}{Character. Term used to perform search.}

\item{user_id}{Integer or character. A user ID or devid provided by Public
Transport Victoria. See section "Obtaining API authentication details".}

\item{api_key}{Character. An API key, with dashes, provided by Public
Transport Victoria. See section "Obtaining API authentication details".}
}
\value{
The response of the \code{search} API call.
}
\description{
There's only one search API call, and it covers stops, routes, and outlets.
This function will return the response of this generic search. It is
expected that other functions will take on of these three categories of
search results, parse them, and return them to the user as a tibble.
}
\details{
If the search term is numeric and/or less than 3 characters, the API will
return only routes.
}
\keyword{internal}
