% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptv-verbs.R
\name{PTVPOST}
\alias{PTVPOST}
\title{Submit a POST request to the PTV API}
\usage{
PTVPOST(request, user_id, api_key, body = NULL, ...)
}
\arguments{
\item{request}{A request or path for the API, eg. "routes".}

\item{user_id}{Integer or character. A user ID or devid provided by Public
Transport Victoria. See section "Obtaining API authentication details".}

\item{api_key}{Character. An API key, with dashes, provided by Public
Transport Victoria. See section "Obtaining API authentication details".}

\item{body}{Body of the POST request. Defaults to NULL, which submits an empty body. Refer to \code{httr::POST} for more details.}

\item{...}{Additional arguments passed to \code{httr::POST}.}
}
\value{
A HTTP response. Content can be accessed with \code{httr::content}.
}
\description{
Submit a POST request to the PTV API
}
\section{Obtaining API authentication details}{


You will need to obtain a user ID (also called a devid) and an API key from
Public Transport Victoria. These are obtained by email. Instructions are
available at
\url{https://www.ptv.vic.gov.au/footer/data-and-reporting/datasets/ptv-timetable-api/}.
}

