% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routes.R
\name{route_types_cached}
\alias{route_types_cached}
\title{Retrieve a translation from route type number to name, cached if possible}
\usage{
route_types_cached(
  user_id = determine_user_id(),
  api_key = determine_api_key()
)
}
\arguments{
\item{user_id}{Integer or character. A user ID or devid provided by Public
Transport Victoria. See section "Obtaining API authentication details".}

\item{api_key}{Character. An API key, with dashes, provided by Public
Transport Victoria. See section "Obtaining API authentication details".}
}
\value{
A named vector in which the values are the route type descriptions,
and the names of the vector are the route type numbers.
}
\description{
This function is equivalent to \code{route_types()}, but will attempt to use a
cached copy of the results stored in options if it is available. If not
available, \code{route_types()} will be called and the results cached. Since
route types are commonly used in other API calls, this caching reduces the
number of times the API is called.
}
\keyword{internal}
