% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routes.R
\name{routes}
\alias{routes}
\title{Retrieve a data frame of all routes}
\usage{
routes(user_id = determine_user_id(), api_key = determine_api_key())
}
\arguments{
\item{user_id}{Integer or character. A user ID or devid provided by Public
Transport Victoria. See section "Obtaining API authentication details".}

\item{api_key}{Character. An API key, with dashes, provided by Public
Transport Victoria. See section "Obtaining API authentication details".}
}
\value{
A tibble of routes, with the following columns:
\itemize{
\item route_id
\item route_gtfs_id
\item route_name
\item route_type
\item route_number
\item service_status
\item service_status_timestamp
}
}
\description{
Retrieve a data frame of all routes
}
\details{
All timestamps returned by this function are in Melbourne time.
}
\examples{
\dontrun{routes()}
}
