% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{run_to_tibble}
\alias{run_to_tibble}
\title{Convert a single run to a tibble}
\usage{
run_to_tibble(run)
}
\arguments{
\item{run}{A run, as a list, returned by the \code{runs} API call.}
}
\value{
A tibble with the following columns: \itemize{
\item \code{run_id} (deprecated, use \code{run_ref} instead)
\item \code{run_ref}
\item \code{route_id}
\item \code{route_type}
\item \code{route_type_description}
\item \code{direction_id}
\item \code{run_sequence}
\item \code{final_stop_id}
\item \code{destination_name}
\item \code{status}
\item \code{express_stop_count}
\item \code{vehicle_position}
\item \code{vehicle_descriptor}
\item \code{geopath}
}
}
\description{
This function is designed to parse the content returned by the interior steps
of the \code{\link{runs_on_route}} and \code{\link{run_information}}
functions.
}
\keyword{internal}
