% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_routes}
\alias{search_routes}
\title{Use a character term to search for routes.}
\usage{
search_routes(
  search_term,
  user_id = determine_user_id(),
  api_key = determine_api_key()
)
}
\arguments{
\item{search_term}{Character. Term used to perform search.}

\item{user_id}{Integer or character. A user ID or devid provided by Public
Transport Victoria. See section "Obtaining API authentication details".}

\item{api_key}{Character. An API key, with dashes, provided by Public
Transport Victoria. See section "Obtaining API authentication details".}
}
\value{
A tibble of routes, with the following columns:
\itemize{
\item route_id
\item route_gtfs_id
\item route_name
\item route_type
\item route_number
\item service_status
\item service_status_timestamp
}
}
\description{
This function will search routes in which the search term can be found in
one of many fields, such as \code{route_id}, \code{route_gtfs_id}, or \code{route_name}.
The search is case-insensitive. Unlike \code{search_stops()} and
\code{search_outlets()}, this function supports searching for numerics, and has
no minimum character requirement for \code{search_term}.
}
\examples{
\dontrun{
search_routes("Pakenham")
}
}
