% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_stops}
\alias{search_stops}
\title{Use a character term to search for stops.}
\usage{
search_stops(
  search_term,
  user_id = determine_user_id(),
  api_key = determine_api_key()
)
}
\arguments{
\item{search_term}{Character. Term used to perform search.}

\item{user_id}{Integer or character. A user ID or devid provided by Public
Transport Victoria. See section "Obtaining API authentication details".}

\item{api_key}{Character. An API key, with dashes, provided by Public
Transport Victoria. See section "Obtaining API authentication details".}
}
\value{
A tibble with the following columns: \itemize{
\item{\code{stop_id}}
\item{\code{stop_name}}
\item{\code{stop_suburb}}
\item{\code{route_type}}
\item{\code{stop_sequence}}
\item{\code{stop_latitude}}
\item{\code{stop_longitude}}
\item{\code{disruption_ids}}
}
}
\description{
This function will search stops in which the search term can be found in
either the stop name or the stop suburb. The search is case-insensitive.
The search term must contain at least 3 characters, and cannot be numeric.
}
\examples{
\dontrun{
search_stops("Ascot Vale")
}
}
