% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops.R
\name{stop_information}
\alias{stop_information}
\title{Stop information (metropolitan and V/Line stations only)}
\usage{
stop_information(
  stop_id,
  route_type,
  user_id = determine_user_id(),
  api_key = determine_api_key()
)
}
\arguments{
\item{stop_id}{Integer. These can be searched for with either \code{stops_on_route} or \code{stops_nearby}.}

\item{route_type}{A route type which can be provided either as a non-negative
integer code, or as a character: "Tram", "Train", "Bus", "Vline" or "Night
Bus". Character inputs are not case-sensitive. Use the \code{route_types}
function to extract a vector of all route types.}

\item{user_id}{Integer or character. A user ID or devid provided by Public
Transport Victoria. See section "Obtaining API authentication details".}

\item{api_key}{Character. An API key, with dashes, provided by Public
Transport Victoria. See section "Obtaining API authentication details".}

\item{route_id}{Integer. These can be listed and described with the \code{routes}
function.}
}
\value{
Stops
}
\description{
Stops can be searched for by exactly one of the following methods: \itemize{
\item Both a \code{stop_id} and \code{route_type} \item Both a \code{route_id} and
\code{route_type} \item Both a \code{latitude} and \code{longitude} }
}
\section{Swagger documentation}{

\url{http://timetableapi.ptv.vic.gov.au/swagger/ui/index#/Stops}
}

