% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops.R
\name{stop_to_tibble}
\alias{stop_to_tibble}
\title{Convert a single stop to a tibble}
\usage{
stop_to_tibble(stop)
}
\arguments{
\item{route}{A stop, as a list, returned by the \code{stops} API call.}
}
\value{
A tibble with the following columns: \itemize{
\item{\code{stop_id}}
\item{\code{stop_name}}
\item{\code{stop_suburb}}
\item{\code{route_type}}
\item{\code{stop_sequence}}
\item{\code{stop_latitude}}
\item{\code{stop_longitude}}
\item{\code{disruption_ids}}
}
}
\description{
This function is designed to parse the content returned by the interior
steps of the \code{stops_on_route} and \code{stop_nearby} functions.
}
\keyword{internal}
