% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops.R
\name{stops_nearby}
\alias{stops_nearby}
\title{Search for stops near a location}
\usage{
stops_nearby(
  latitude,
  longitude,
  route_types = NULL,
  user_id = determine_user_id(),
  api_key = determine_api_key()
)
}
\arguments{
\item{latitude}{Numeric. Latitude in decimal degrees. For example, Flinders
Street Station is at approximately -37.8183 latitude.}

\item{longitude}{Numeric. Longitude in decimal degrees. For example, Flinders
Street Station is at approximately 144.9671 longitude.}

\item{route_types}{Optionally filter by a vector of route types. A route type
can be provided either as a non-negative integer code, or as a character:
"Tram", "Train", "Bus", "Vline" or "Night Bus". Character inputs are not
case-sensitive. Use the \code{route_types} function to extract a vector of all
route types.}

\item{user_id}{Integer or character. A user ID or devid provided by Public
Transport Victoria. See section "Obtaining API authentication details".}

\item{api_key}{Character. An API key, with dashes, provided by Public
Transport Victoria. See section "Obtaining API authentication details".}
}
\value{
A tibble with the following columns: \itemize{
\item{\code{stop_id}}
\item{\code{stop_name}}
\item{\code{stop_suburb}}
\item{\code{route_type}}
\item{\code{stop_sequence}}
\item{\code{stop_latitude}}
\item{\code{stop_longitude}}
\item{\code{disruption_ids}}
}
}
\description{
Search for stops near a location
}
\section{Swagger documentation}{

\url{http://timetableapi.ptv.vic.gov.au/swagger/ui/index#/Stops}
}

\examples{
\dontrun{
stops_nearby(latitude = -37.8183, longitude = 144.9671)
}
}
