% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops.R
\name{stops_with_stop_id}
\alias{stops_with_stop_id}
\title{Stop information with stop_id and route_type}
\usage{
stops_with_stop_id(
  stop_id,
  route_type,
  user_id = determine_user_id(),
  api_key = determine_api_key()
)
}
\arguments{
\item{stop_id}{TODO}

\item{route_type}{TODO}

\item{user_id}{Integer or character. A user ID or devid provided by Public
Transport Victoria. See section "Obtaining API authentication details".}

\item{api_key}{Character. An API key, with dashes, provided by Public
Transport Victoria. See section "Obtaining API authentication details".}
}
\value{
Stops
}
\description{
Stop information with stop_id and route_type
}
\section{Swagger documentation}{

\url{http://timetableapi.ptv.vic.gov.au/swagger/ui/index#/Stops}
}

\keyword{internal}
