% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{to_datetime}
\alias{to_datetime}
\title{Convert a POSIXct or character datetime to a format ready for a URL}
\usage{
to_datetime(datetime)
}
\arguments{
\item{datetime}{POSIXct or Character.}
}
\value{
Character.
}
\description{
Datetimes accepted by the API need to be given in UTC. This function will
accept a datetime or a character with a suitable datetime format, and output
a character that is suitable for a URL. All URL input and output in this
package should be in Melbourne time, and the UTC conversion should happen
}
\details{
The API seems to accept both "\%Y-\%m-\%dT\%H:\%M:\%OS" and "\%Y-\%m-\%d \%H:\%M:\%OS", but we opt for the former.
}
\keyword{internal}
