% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routes.R
\name{translate_route_types}
\alias{translate_route_types}
\title{Translate a route type input into a numerical route type}
\usage{
translate_route_types(route_type)
}
\arguments{
\item{route_type}{A route type which can be provided either as a non-negative
integer code, or as a character: "Tram", "Train", "Bus", "Vline" or "Night
Bus". Character inputs are not case-sensitive. Use the \code{route_types}
function to extract a vector of all route types.}
}
\value{
An integer route type code
}
\description{
Many API calls require a route type (eg. "Tram" or "Train"). These must be
provided as integers, which are translated to route type descriptions with
}
\keyword{internal}
