% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{user_id}
\alias{user_id}
\title{Return the user ID from the PTV_USER_ID environment variable}
\usage{
user_id()
}
\value{
Character. The value of the PTV_USER_ID environment variable.
}
\description{
Currently this package only supports obtaining the user ID either directly
as an argument to functions, or from the PTV_USER_ID environment variable.
This function will error if the PTV_USER_ID environment variable is not set.
A message will also be displayed to the user, explaining how to obtain
authentication details from PTV. This message will be displayed only once
per session.
}
\keyword{internal}
