% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directions.R
\name{directions}
\alias{directions}
\title{Directions for a given route}
\usage{
directions(
  route_id,
  user_id = determine_user_id(),
  api_key = determine_api_key()
)
}
\arguments{
\item{route_id}{Integer. These can be listed and described with the \code{routes}
function.}

\item{user_id}{Integer or character. A user ID or devid provided by Public
Transport Victoria. See section "Obtaining API authentication details".}

\item{api_key}{Character. An API key, with dashes, provided by Public
Transport Victoria. See section "Obtaining API authentication details".}
}
\value{
directions
}
\description{
Directions for a given route
}
