% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-request-url.R
\name{prefix_base_url_and_version}
\alias{prefix_base_url_and_version}
\title{Prefix a request with the base URL and version}
\usage{
prefix_base_url_and_version(request)
}
\arguments{
\item{request}{A request or path for the API, eg. "routes".}
}
\value{
A complete URL character
}
\description{
Prefix a request with the base URL and version
}
\keyword{internal}
