% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routes.R
\name{route_to_tibble}
\alias{route_to_tibble}
\title{Convert a single route to a tibble}
\usage{
route_to_tibble(route)
}
\arguments{
\item{route}{A route, as a list, returned by the \code{routes} API call.}
}
\value{
A tibble of routes, with the following columns:
\itemize{
\item \code{route_id}
\item \code{route_gtfs_id}
\item \code{route_name}
\item \code{route_type}
\item \code{route_type_description}
\item \code{route_number}
\item \code{geopath}
\item \code{service_status}
\item \code{service_status_timestamp}
}
}
\description{
This function is designed to parse the content returned by the interior
steps of the \code{routes} function. Service status information may be \code{NA},
depending on the API call that was used to populate the information.
}
\keyword{internal}
