% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routes.R
\name{routes}
\alias{routes}
\title{Retrieve route information}
\usage{
routes(
  route_types = NULL,
  route_name = NULL,
  user_id = determine_user_id(),
  api_key = determine_api_key()
)
}
\arguments{
\item{route_types}{Integer or character vector. Optionally filter by a vector
of route types. A route type can be provided either as a non-negative
integer code, or as a character: "Tram", "Train", "Bus", "Vline" or "Night
Bus". Character inputs are not case-sensitive. Use the \code{route_types}
function to extract a vector of all route types.}

\item{route_name}{Character. Optionally filter by route name. Partial matches
are accepted, and the matches are not case sensitive.}

\item{user_id}{Integer or character. A user ID or devid provided by Public
Transport Victoria. See section "Obtaining API authentication details".}

\item{api_key}{Character. An API key, with dashes, provided by Public
Transport Victoria. See section "Obtaining API authentication details".}
}
\value{
A tibble of routes, with the following columns:
\itemize{
\item route_id
\item route_gtfs_id
\item route_name
\item route_type
\item route_number
\item service_status
\item service_status_timestamp
}
}
\description{
Retrieve route information
}
\details{
All timestamps returned by this function are in Melbourne time.
}
\examples{
\dontrun{routes()}
}
