% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{run_information}
\alias{run_information}
\title{Retrieve a tibble of all runs on a particular route}
\usage{
run_information(
  run_id,
  route_type = NULL,
  user_id = determine_user_id(),
  api_key = determine_api_key()
)
}
\arguments{
\item{run_id}{An integer run ID. This may retrieved from the \code{departures()}
or \code{runs_on_route()} functions.}

\item{route_type}{Optionally filter results by a route type. A route type can
be provided either as a non-negative integer code, or as a character:
"Tram", "Train", "Bus", "Vline" or "Night Bus". Character inputs are not
case-sensitive. Use the \code{route_types} function to extract a vector of all
route types.}

\item{user_id}{Integer or character. A user ID or devid provided by Public
Transport Victoria. See section "Obtaining API authentication details".}

\item{api_key}{Character. An API key, with dashes, provided by Public
Transport Victoria. See section "Obtaining API authentication details".}
}
\value{
A tibble with the following columns: \itemize{
\item \code{run_id}
\item \code{route_id}
\item \code{route_type}
\item \code{direction_id}
\item \code{run_sequence}
\item \code{final_stop_id}
\item \code{destination_name}
\item \code{status}
\item \code{express_stop_count}
\item \code{vehicle_position}
\item \code{vehicle_descriptor}
}
}
\description{
Run IDs are not unique across the network. If you are interested in a
specific run, consider supplying a value to the optional \code{route_type}
argument.
}
