% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops.R
\name{stops}
\alias{stops}
\title{Stop information (metropolitan and V/Line stations only)}
\usage{
stops(
  stop_id,
  route_id,
  route_type,
  latitude,
  longitude,
  user_id = determine_user_id(),
  api_key = determine_api_key()
)
}
\arguments{
\item{api_key}{}
}
\value{
Stops
}
\description{
Stops can be searched for by exactly one of the following methods:
\itemize{
\item Both a \code{stop_id} and \code{route_type}
\item Both a \code{route_id} and \code{route_type}
\item Both a \code{latitude} and \code{longitude}
}
}
\section{Swagger documentation}{

\url{http://timetableapi.ptv.vic.gov.au/swagger/ui/index#/Stops}
}

